Function server_config::load_config

pub fn load_config(data: Option<ConfigData>) -> Result<UnresolvedConfig>

Hierarchically loads the configuration using args and envs. The source order is:

Load and parse Config.toml from cwd, if exists

Load and parse files provided by FLUENCE_CONFIG env var

Load and parse files provided by –config arg

Load config values from env vars

Load config values from args (throw error on conflicts with env vars) On each stage the values override the previous ones.

Arguments

data: Optional ConfigData to customize the configuration.

Returns

Returns a Result containing the unresolved configuration or an Eyre error.