Struct server_config::ServicesConfig

pub struct ServicesConfig {

    pub local_peer_id: PeerId,
    pub blueprint_dir: PathBuf,
    pub envs: HashMap<String, String>,
    pub persistent_work_dir: PathBuf,
    pub ephemeral_work_dir: PathBuf,
    pub modules_dir: PathBuf,
    pub services_dir: PathBuf,
    pub particles_vault_dir: PathBuf,
    pub management_peer_id: PeerId,
    pub builtins_management_peer_id: PeerId,
    pub default_service_memory_limit: Option<ByteSize>,
    pub allowed_effectors: HashMap<Hash, HashMap<String, PathBuf>>,
    pub mounted_binaries_mapping: HashMap<String, PathBuf>,
    pub is_dev_mode: bool,

}

Fields
local_peer_id: PeerId

Peer id of the current node

blueprint_dir: PathBuf

Path of the blueprint directory containing blueprints and wasm modules

envs: HashMap<String, String>

Opaque environment variables to be passed on each service creation TODO: isolate envs of different modules (i.e., module A shouldn’t access envs of module B)

persistent_work_dir: PathBuf

Persistent working dir for services

ephemeral_work_dir: PathBuf

Ephemeral working dir for services

modules_dir: PathBuf

Dir to store .wasm modules and their configs

services_dir: PathBuf

Dir to persist info about running services

particles_vault_dir: PathBuf

Dir to store directories shared between services in the span of a single particle execution

management_peer_id: PeerId

key that could manage services

builtins_management_peer_id: PeerId

key to manage builtins services initialization

default_service_memory_limit: Option<ByteSize>

Default heap size in bytes available for the module unless otherwise specified.

allowed_effectors: HashMap<Hash, HashMap<String, PathBuf>>

List of allowed effector modules by CID

mounted_binaries_mapping: HashMap<String, PathBuf>

Mapping of binary names to their paths for mounted binaries used in developer mode

is_dev_mode: bool

Is in the developer mode

Implementations

impl ServicesConfig

pub fn new( local_peer_id: PeerId, persistent_dir: PathBuf, ephemeral_dir: PathBuf, particles_vault_dir: PathBuf, envs: HashMap<String, String>, management_peer_id: PeerId, builtins_management_peer_id: PeerId, default_service_memory_limit: Option<ByteSize>, allowed_effectors: HashMap<Hash, HashMap<String, String>>, mounted_binaries_mapping: HashMap<String, String>, is_dev_mode: bool ) -> Result<Self, Error>